@@ -16,31 +16,31 @@
 		"npm": ">=9.6.7"
 	},
 	"devDependencies": {
-		"@sveltejs/adapter-auto": "^3.0.0",
-		"@sveltejs/adapter-cloudflare": "^4.1.0",
-		"@sveltejs/kit": "^2.0.0",
-		"@sveltejs/vite-plugin-svelte": "^3.0.0",
+		"@sveltejs/adapter-auto": "^3.3.1",
+		"@sveltejs/adapter-cloudflare": "^4.9.0",
+		"@sveltejs/kit": "^2.49.4",
+		"@sveltejs/vite-plugin-svelte": "^5.1.1",
 		"@tailwindcss/typography": "^0.5.16",
 		"@types/eslint": "^9.6.0",
 		"autoprefixer": "^10.4.20",
 		"eslint": "^9.0.0",
 		"eslint-config-prettier": "^9.1.0",
-		"eslint-plugin-svelte": "^2.36.0",
+		"eslint-plugin-svelte": "^2.45.1",
 		"globals": "^15.0.0",
 		"prettier": "^3.1.1",
-		"prettier-plugin-svelte": "^3.1.2",
+		"prettier-plugin-svelte": "^3.2.6",
 		"prettier-plugin-tailwindcss": "^0.6.5",
-		"svelte": "^4.2.7",
-		"svelte-check": "^3.6.0",
+		"svelte": "^5.33.0",
+		"svelte-check": "^4.3.5",
 		"tailwindcss": "^3.4.9",
-		"typescript": "^5.0.0",
+		"typescript": "^5.5.0",
 		"typescript-eslint": "^8.0.0",
-		"vite": "^5.0.3"
+		"vite": "^6.3.0"
 	},
 	"type": "module",
 	"dependencies": {
 		"@fortawesome/fontawesome-free": "^6.6.0",
-		"bits-ui": "^0.21.13",
+		"bits-ui": "^2.15.4",
 		"clsx": "^2.1.1",
 		"markdown-it": "^14.1.0",
 		"svelte-inview": "^4.0.4",
